Inflate dialog with layout resource so that theming is correct
